Create QuickSort.java
If the array contains more than one element than:

Select an element from the array. This element is called the “pivot element”. For example select the element in the middle of the array.
All elements which are smaller then the pivot element are placed in one array and all elements which are larger are placed in another array.
Sort both arrays by recursively applying Quicksort to them.
Combine the arrays.
Quicksort can be implemented to sort “in-place”. This means that the sorting takes place in the array and that no additional array needs to be created.

This is a Java Program to perform Quick Sort Algorithm.